Environment variables (#156)
* Rename NEXT_PUBLIC_NETWORK to NETWORK, add NEXT_PUBLIC_BASE_API_URL


* UI image always builds


* Do not build UI image in Actions


* Code tidy

+++ hedera-nft-auction-demo-java-node/src/main/java/com/hedera/demo/auction/app/EasySetup.java
@@ -38,9 +38,15 @@ public class EasySetup extends AbstractCreate {
 
         if (requestEasySetup.clean) {
             log.info("Deleting existing auctions and bids and creating new topic");
-            bidsRepository.deleteAllBids();
-            auctionsRepository.deleteAllAuctions();
-            validatorsRepository.deleteAllValidators();
+            if (bidsRepository != null) {
+                bidsRepository.deleteAllBids();
+            }
+            if (auctionsRepository != null) {
+                auctionsRepository.deleteAllAuctions();
+            }
+            if (validatorsRepository != null) {
+                validatorsRepository.deleteAllValidators();
+            }
             CreateTopic createTopic = new CreateTopic();
             topicId = createTopic.create().toString();
         }
